+++ CHANGELOG.md
@@ -8,6 +8,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-01-01
+
+### Added
+
+* In addition to the `HTML-like` format, add the `JSON` format for
+  exporting all entries from the database and importing data to
+  the database.
+
 ## [1.0.1] - 2025-12-31
 
 ### Fixed